source/requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
docker==6.1.3
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
packaging==23.1
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
websocket-client==1.6.2
